day-20-snake-game/requirements.txt
openai>=1.0.0
anthropic>=0.34.0
python-dotenv>=1.0.0 